api_yamdb/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
django-filter==22.1
tqdm==4.64.1
asgiref==3.6.0
sqlparse==0.4.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
